fix MDI-X showing as Unknown instead of "off (auto)"
The current version of ethtool is unable to show the correct MDI-X info:
 # ethtool --version
 ethtool version 6.11
 # ethtool enp0s31f6 | grep MDI
     MDI-X: Unknown

For comparison, an older version shows it correctly:
 # ./ethtool --version
 ethtool version 6.2
 # ./ethtool enp0s31f6 | grep MDI
     MDI-X: off (auto)

The blamed commit accidentally removed the ETH_TP_MDI switch case
in dump_mdix(). As a result, ETH_TP_MDI is treated like Unknown.

Fix it by restoring the ETH_TP_MDI case and breaking out.
'mdi_x' is initialized to false, which is correct for this case.

Fixes: bd1341cd2146 ("add json support for base command")
Signed-off-by: Michal Schmidt <mschmidt@redhat.com>