Print unknown RSS hash function value when encountered
It is possible for the kernel to report unknown RSS hash function, for
example, when a new driver which uses a new hash function by default is
used with an old userspace ethtool.

When such hash function is encountered, print its raw value so the user
is at least aware that some unparsable function is used.

Reviewed-by: Cosmin Ratiu <cratiu@nvidia.com>
Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: Gal Pressman <gal@nvidia.com>