Print unknown input_xfrm values when encountered
It is possible for the kernel to report unknown input_xfrm values, e.g.
when a new driver which enables a new transformation by default is used
with an old userspace ethtool.

When such values are encountered, print the raw values so the user is at
least aware that some unparsable transformations are enabled.

Reviewed-by: Cosmin Ratiu <cratiu@nvidia.com>
Reviewed-by: Tariq Toukan <tariqt@nvidia.com>
Signed-off-by: Gal Pressman <gal@nvidia.com>